requirements.txt
streamlit
opencv-python
numpy
pandas
Pillow
imutils
scikit-image
openpyxl
